requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# LLM & Agent Framework (Latest Langchain)
langchain==0.3.14
langchain-openai==0.2.14
langchain-community==0.3.14
langchain-core==0.3.29

# OpenAI API (for embeddings and LLM)
openai==1.59.5

# HTTP Requests for APIs
requests==2.31.0

# Document Processing
pdfplumber
python-docx

# Vector Store & Embeddings
chromadb==0.4.22
numpy==1.26.3

# Web Framework
flask==3.0.0

# Testing
pytest
